chore(frontend): :truck: Move bot-panel to its own directory
