--- frontend/src/bot-summary-panel/input.interface.ts
@@ -1,10 +1,0 @@
-import { Bot, Position } from '../rpc/bot_pb';
-
-export interface Input {
-  /** Bot summary */
-  bot: Bot,
-  /** Current Positions */
-  curPos: Position[],
-  /** Archived Positions */
-  arcPos: Position[],
-}
