--- frontend/src/bot-summary-panel/view.tsx
@@ -1,1 +1,0 @@
-export default function BotSummaryPanel() { }
